Expose IO
