+++ divik/core/io/__init__.py
@@ -1,0 +1,9 @@
+from ._data_io import (
+    load_data,
+)
+from ._model_io import (
+    saver,
+    save
+)
+
+DIVIK_RESULT_FNAME = 'result.pkl'
